+++ src/lib.rs
@@ -105,7 +105,7 @@ pub unsafe trait WriteBuffer {
 
 unsafe impl<B, T> ReadBuffer for B
 where
-    B: Deref<Target = T> + StableDeref,
+    B: Deref<Target = T> + StableDeref + 'static,
     T: ReadTarget + ?Sized,
 {
     type Word = T::Word;
@@ -117,7 +117,7 @@ where
 
 unsafe impl<B, T> WriteBuffer for B
 where
-    B: DerefMut<Target = T> + StableDeref,
+    B: DerefMut<Target = T> + StableDeref + 'static,
     T: WriteTarget + ?Sized,
 {
     type Word = T::Word;
@@ -153,7 +153,7 @@ unsafe impl Word for i64 {}
 /// # Safety
 ///
 /// - `as_read_buffer` must adhere to the safety requirements
-///   documented for `DmaReadBuffer::dma_read_buffer`.
+///   documented for `ReadBuffer::read_buffer`.
 pub unsafe trait ReadTarget {
     type Word: Word;
 
@@ -172,7 +172,7 @@ pub unsafe trait ReadTarget {
 /// # Safety
 ///
 /// - `as_write_buffer` must adhere to the safety requirements
-///   documented for `DmaWriteBuffer::dma_write_buffer`.
+///   documented for `WriteBuffer::write_buffer`.
 pub unsafe trait WriteTarget {
     type Word: Word;
 
@@ -199,130 +199,16 @@ unsafe impl<T: WriteTarget> WriteTarget for [T] {
     type Word = T::Word;
 }
 
-macro_rules! dma_target_array_impls {
-    ( $( $i:expr, )+ ) => {
-        $(
-            unsafe impl<T: ReadTarget> ReadTarget for [T; $i] {
-                type Word = T::Word;
-            }
-
-            unsafe impl<T: WriteTarget> WriteTarget for [T; $i] {
-                type Word = T::Word;
-            }
-        )+
-    };
-}
-
-#[rustfmt::skip]
-dma_target_array_impls!(
-     0,   1,   2,   3,   4,   5,   6,   7,   8,   9,
-    10,  11,  12,  13,  14,  15,  16,  17,  18,  19,
-    20,  21,  22,  23,  24,  25,  26,  27,  28,  29,
-    30,  31,  32,  33,  34,  35,  36,  37,  38,  39,
-    40,  41,  42,  43,  44,  45,  46,  47,  48,  49,
-    50,  51,  52,  53,  54,  55,  56,  57,  58,  59,
-    60,  61,  62,  63,  64,  65,  66,  67,  68,  69,
-    70,  71,  72,  73,  74,  75,  76,  77,  78,  79,
-    80,  81,  82,  83,  84,  85,  86,  87,  88,  89,
-    90,  91,  92,  93,  94,  95,  96,  97,  98,  99,
-   100, 101, 102, 103, 104, 105, 106, 107, 108, 109,
-   110, 111, 112, 113, 114, 115, 116, 117, 118, 119,
-   120, 121, 122, 123, 124, 125, 126, 127, 128, 129,
-   130, 131, 132, 133, 134, 135, 136, 137, 138, 139,
-   140, 141, 142, 143, 144, 145, 146, 147, 148, 149,
-   150, 151, 152, 153, 154, 155, 156, 157, 158, 159,
-   160, 161, 162, 163, 164, 165, 166, 167, 168, 169,
-   170, 171, 172, 173, 174, 175, 176, 177, 178, 179,
-   180, 181, 182, 183, 184, 185, 186, 187, 188, 189,
-   190, 191, 192, 193, 194, 195, 196, 197, 198, 199,
-   200, 201, 202, 203, 204, 205, 206, 207, 208, 209,
-   210, 211, 212, 213, 214, 215, 216, 217, 218, 219,
-   220, 221, 222, 223, 224, 225, 226, 227, 228, 229,
-   230, 231, 232, 233, 234, 235, 236, 237, 238, 239,
-   240, 241, 242, 243, 244, 245, 246, 247, 248, 249,
-   250, 251, 252, 253, 254, 255, 256,
-
-   1 <<  9,
-   1 << 10,
-   1 << 11,
-   1 << 12,
-   1 << 13,
-   1 << 14,
-   1 << 15,
-   1 << 16,
-);
-
-unsafe impl<T: WriteTarget> WriteTarget for MaybeUninit<T> {
+unsafe impl<T: ReadTarget, const N: usize> ReadTarget for [T; N] {
     type Word = T::Word;
 }
 
-/// Trait for buffers that can be given to DMA for reading. This is a more strict version of
-/// [ReadBuffer](trait.ReadBuffer.html), if you are not sure about which one to use on your safe
-/// API, prefer this one. This trait also allows end users to __unsafely__ bypass the `'static`
-/// invariant.
-///
-/// # Safety
-///
-/// This has the same invariants as [ReadBuffer](trait.ReadBuffer.html) with the additional
-/// requirement that the buffer should have a `'static` lifetime.
-pub unsafe trait StaticReadBuffer: ReadBuffer {
-    type Word;
-
-    /// Provide a buffer usable for DMA reads.
-    ///
-    /// The return value is:
-    ///
-    /// - pointer to the start of the buffer
-    /// - buffer size in words
-    ///
-    /// # Safety
-    ///
-    /// Once this method has been called, it is unsafe to call any `&mut self`
-    /// methods on this object as long as the returned value is in use (by DMA).
-    unsafe fn static_read_buffer(&self) -> (*const <Self as StaticReadBuffer>::Word, usize);
+unsafe impl<T: WriteTarget, const N: usize> WriteTarget for [T; N] {
+    type Word = T::Word;
 }
 
-/// Trait for buffers that can be given to DMA for writing. This is a more strict version of
-/// [WriteBuffer](trait.WriteBuffer.html), if you are not sure about which one to use on your safe
-/// API, prefer this one. This trait also allows end users to __unsafely__ bypass the `'static`
-/// invariant.
-///
-/// # Safety
-///
-/// This has the same invariants as [WriteBuffer](trait.WriteBuffer.html) with the additional
-/// requirement that the buffer should have a `'static` lifetime.
-pub unsafe trait StaticWriteBuffer: WriteBuffer {
-    type Word;
-
-    /// Provide a buffer usable for DMA writes.
-    ///
-    /// The return value is:
-    ///
-    /// - pointer to the start of the buffer
-    /// - buffer size in words
-    ///
-    /// # Safety
-    ///
-    /// Once this method has been called, it is unsafe to call any `&mut self`
-    /// methods, except for `write_buffer`, on this object as long as the
-    /// returned value is in use (by DMA).
-    unsafe fn static_write_buffer(&mut self) -> (*mut <Self as StaticWriteBuffer>::Word, usize);
-}
-
-unsafe impl<B: ReadBuffer + 'static> StaticReadBuffer for B {
-    type Word = <Self as ReadBuffer>::Word;
-
-    unsafe fn static_read_buffer(&self) -> (*const <Self as StaticReadBuffer>::Word, usize) {
-        self.read_buffer()
-    }
-}
-
-unsafe impl<B: WriteBuffer + 'static> StaticWriteBuffer for B {
-    type Word = <Self as WriteBuffer>::Word;
-
-    unsafe fn static_write_buffer(&mut self) -> (*mut <Self as StaticWriteBuffer>::Word, usize) {
-        self.write_buffer()
-    }
+unsafe impl<T: WriteTarget> WriteTarget for MaybeUninit<T> {
+    type Word = T::Word;
 }
 
 #[cfg(test)]
@@ -337,25 +223,11 @@ mod tests {
         unsafe { buffer.read_buffer() }
     }
 
-    fn static_api_read<W, B>(buffer: B) -> (*const W, usize)
-    where
-        B: StaticReadBuffer<Word = W>,
-    {
-        unsafe { buffer.static_read_buffer() }
-    }
-
     fn api_write<W, B>(mut buffer: B) -> (*mut W, usize)
     where
         B: WriteBuffer<Word = W>,
     {
         unsafe { buffer.write_buffer() }
-    }
-
-    fn static_api_write<W, B>(mut buffer: B) -> (*mut W, usize)
-    where
-        B: StaticWriteBuffer<Word = W>,
-    {
-        unsafe { buffer.static_write_buffer() }
     }
 
     #[test]
@@ -364,13 +236,9 @@ mod tests {
         static BUF: [u8; SIZE] = [0u8; SIZE];
         let local_buf = [0u8; SIZE];
 
-        let (ptr, size_local) = api_read(&local_buf);
+        let (ptr, size_local) = api_read(&BUF);
         assert!(unsafe { (&*ptr as &dyn Any).is::<u8>() });
         assert!(size_local == SIZE);
-
-        let (ptr, size_static) = static_api_read(&BUF);
-        assert!(unsafe { (&*ptr as &dyn Any).is::<u8>() });
-        assert!(size_static == SIZE);
     }
 
     #[test]
@@ -379,12 +247,8 @@ mod tests {
         static mut BUF: [u8; SIZE] = [0u8; SIZE];
         let mut local_buf = [0u8; SIZE];
 
-        let (ptr, size_local) = api_write(&mut local_buf);
+        let (ptr, size_local) = api_write(unsafe { &mut BUF });
         assert!(unsafe { (&*ptr as &dyn Any).is::<u8>() });
         assert!(size_local == SIZE);
-
-        let (ptr, size_static) = static_api_write(unsafe { &mut BUF });
-        assert!(unsafe { (&*ptr as &dyn Any).is::<u8>() });
-        assert!(size_static == SIZE);
     }
 }